Refactored image.py to support image apis for different rendering backends. load_pixels is no longer a context manager as it was different than p5.js and update_pixel is implemented
